requirements-dev.txt
# Include production dependencies
-r bot/requirements.txt

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Type stubs
types-cachetools==5.3.0.7
sqlalchemy-stubs==0.4
